pyproject.toml
[project]
name = 'polymarket-bot'
version = '0.1.0'
requires-python = '>=3.11'
dependencies = [
  'httpx>=0.27',
  'pydantic>=2.6',
  'pandas>=2.2',
  'python-dotenv>=1.0',
]

[project.optional-dependencies]
dev = [
  'pytest>=8.0',
]

[project.scripts]
polymarket = "polymarket.cli:main"

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v"
